RU-Parking-master/RU-Parking-master/requirements.txt
flask
flask-cors
openai==1.40.0
httpx==0.27.0
httpcore==1.0.4
jiter==0.12.0
python-dotenv
